refactor(spectacular): remove barrels
