+++ packages/spectacular/src/index.ts
@@ -1,2 +1,7 @@
+// Feature harness
 export * from './lib/create-feature-test-harness';
-export * from './lib/pipe-test-harness';
+
+// Pipe test harness
+export * from './lib/pipe-test-harness/create-pipe-test-harness';
+export * from './lib/pipe-test-harness/create-pipe-test-harness-options';
+export * from './lib/pipe-test-harness/pipe-test-harness';
