--- packages/spectacular/src/lib/pipe-test-harness/index.ts
@@ -1,3 +1,0 @@
-export * from './create-pipe-test-harness';
-export * from './create-pipe-test-harness-options';
-export * from './pipe-test-harness';
