--- packages/spectacular/src/lib/pipe-test-harness/test-pipe/index.ts
@@ -1,1 +1,0 @@
-export * from './test-pipe.component';
