+++ packages/spectacular/src/lib/pipe-test-harness/test-util/create-common-pipe-test-harness-options.ts
@@ -1,4 +1,5 @@
 import type { Observable } from 'rxjs';
+
 export interface CreateCommonPipeTestHarnessOptions<TValue> {
   readonly template: string;
   readonly value: Observable<TValue> | TValue | null;
