--- packages/spectacular/src/lib/pipe-test-harness/test-util/index.ts
@@ -1,2 +1,0 @@
-export * from './create-common-pipe-test-harness';
-export * from './create-common-pipe-test-harness-options';
