--- packages/spectacular/src/lib/test-root/index.ts
@@ -1,2 +1,0 @@
-export * from './test-root.component';
-export * from './test-root.scam';
