--- packages/spectacular/src/lib/util/index.ts
@@ -1,2 +1,0 @@
-export * from './create-left-mouse-click';
-export * from './text-utilities';
